Parse and sanitize URL before sending it to open::that
