+++ src/stremio_app/constants.rs
@@ -1,3 +1,5 @@
+use url::Url;
+
 pub const APP_NAME: &str = "Stremio";
 pub const IPC_PATH: &str = "//./pipe/com.stremio5.";
 pub const DEV_ENDPOINT: &str = "http://127.0.0.1:11470";
@@ -16,6 +18,32 @@ pub const SRV_BUFFER_SIZE: usize = 1024;
 pub const SERVER_IPC_KEY: &str = "SERVER_IPC_KEY";
 pub const SRV_LOG_SIZE: usize = 20;
 
+pub const WARNING_URL: &str = "https://www.stremio.com/warning#";
+pub const WHITELISTED_HOSTS: &[&str] = &[
+    "stremio.com",
+    "www.stremio.com",
+    "web.stremio.com",
+    "app.stremio.com",
+    "strem.io",
+    "api.strem.io",
+    "stremio.zendesk.com",
+    "google.com",
+    "www.google.com",
+    "youtube.com",
+    "www.youtube.com",
+    "twitch.tv",
+    "twitter.com",
+    "x.com",
+    "netflix.com",
+    "adex.network",
+    "amazon.com",
+    "forms.gle",
+    "www.hbomax.com",
+    "play.hbomax.com",
+    "www.disneyplus.com",
+    "imdb.com",
+];
+
 pub fn web_endpoint_with_streaming_server(server_url: &str) -> String {
     let server_url = server_url.trim_end_matches('/');
     let streaming_server_url =
@@ -23,4 +51,24 @@ pub fn web_endpoint_with_streaming_server(server_url: &str) -> String {
     let web_endpoint = WEB_ENDPOINT.trim_end_matches('/');
 
     format!("{web_endpoint}/#/?streamingServerUrl={streaming_server_url}")
+}
+
+pub fn safe_url(uri: &str) -> Option<String> {
+    if let Ok(url) = Url::parse(uri) {
+        println!("URL is {url}");
+        let is_whitelisted = url.host().is_some_and(|host| {
+            WHITELISTED_HOSTS
+                .iter()
+                .any(|whitelisted_host| host.to_string() == *whitelisted_host)
+        });
+
+        let final_url = if is_whitelisted {
+            url.to_string()
+        } else {
+            format!("{}{}", WARNING_URL, urlencoding::encode(url.as_ref()))
+        };
+        Some(final_url)
+    } else {
+        None
+    }
 }
