--- src/stremio_app/stremio_wevbiew/constants.rs
@@ -1,15 +1,0 @@
-pub const WARNING_URL: &str = "https://www.stremio.com/warning#";
-pub const WHITELISTED_HOSTS: &[&str] = &[
-    "stremio.com",
-    "strem.io",
-    "stremio.zendesk.com",
-    "google.com",
-    "youtube.com",
-    "twitch.tv",
-    "twitter.com",
-    "x.com",
-    "netflix.com",
-    "adex.network",
-    "amazon.com",
-    "forms.gle",
-];
