+++ src/stremio_app/stremio_wevbiew/mod.rs
@@ -1,4 +1,2 @@
-mod constants;
-
 pub mod wevbiew;
 pub use wevbiew::WebView;
